'feat. LV1, LV2 구현, LV3 일부 예외처리'

+++ src/main/java/com/sparta/currency_user/dto/CurrencyRequestDto.java
@@ -3,12 +3,11 @@ package com.sparta.currency_user.dto;
 import com.sparta.currency_user.entity.Currency;
 import lombok.Getter;
 
-import java.math.BigDecimal;
 
 @Getter
 public class CurrencyRequestDto {
     private String currencyName;
-    private BigDecimal exchangeRate;
+    private Double exchangeRate;
     private String symbol;
 
     public Currency toEntity() {
